Add a gnome-settings-daemon-common package
This contains the GSettings schemas and other arch-independent files.
Packages can depend or build-depend on this if they want the schemas but
not the actual running gnome-settings-daemon.

In particular, we currently have a circular BD/D chain between mutter
and g-s-d, and this can be mostly broken if mutter depends on the
schemas (which it needs) rather than the daemon (which it doesn't).

The references to gnome-settings-daemon-schemas are because Ubuntu
provides this package under that name - this is to help with
transitioning to gnome-settings-daemon-common.