debian/control: Add Breaks on old gtk and mutter
Since the font-settings related schemas are now moved to
gsettings-desktop-schemas, this version will break with various
libraries and applications relying on them.

So, add breaks on old versions of gtk3, gtk4 and mutter libraries
(using unreleased versions, as it's not expected that any new version of
them will be adapted to this change).
Do not break on gnome-tweaks, as it will still work even though without
showing the font settings, but this is not an huge requirement.

See: https://gitlab.gnome.org/GNOME/gsettings-desktop-schemas/-/issues/31